requirements.txt
schedule
requests
PyYAML
pandas
pandas-ta
python-dotenv
numpy<2.0 